interim commit with stable "normals + lighting" integration test in waveform_terrain_3d_dc.c and playback control foundation in sound_envelope_3d_audio_cadence_dc.c

+++ src/cool_gl33.c
@@ -25,7 +25,7 @@ int main(void) {
     AudioStream audio_stream = LoadAudioStream(SRC_SAMPLE_RATE, SRC_BIT_DEPTH, SRC_CHANNELS);
 
     PlayAudioStream(audio_stream);
-    size_t wave_cursor = 0;
+    unsigned int wave_cursor = 0;
     int16_t* wave_pcm16 = (int16_t*)wave.data;
     Image fft_image = GenImageColor(ANALYSIS_SPECTRUM_BIN_COUNT, 1, BLACK);
     Color* fft_pixels = (Color*)fft_image.data;
